refactor: resposne class to static inner class
